[vm, gc] Reserve ForwardingPages up front.
Allocate space for ForwardingPages when allocating a new HeapPages that is subject to compaction, instead of allocating it at compaction time, which is when we're already low on memory.

This has a 4KB/256KB = 1.5% overhead for ordinary pages, no overhead for large and code pages.

Bug: https://github.com/dart-lang/sdk/issues/37772
Change-Id: I9ce4d0791093df8ee98d9bda17f0a1cc33788e28
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/114020
Commit-Queue: Ryan Macnak <rmacnak@google.com>
Reviewed-by: Siva Annamalai <asiva@google.com>