Make exitFunctionBody safer.
There's no functional change here; it was wrong to have mismatched
exit/enter of a function body before, and it's still wrong.  But this
change makes it easier to track down problems if an exception gets
thrown debug visiting, causing a mismatch; if that happens we can
temporarily comment out the assert to avoid covering up the exception.

Change-Id: Iad1e820f65387b4b2a0537d763b887df18d7fbf4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/120480
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>