Stop using lazy FunctionType where possible.
We have to keep it for now for GenericFunctionType nodes, because
of the way we build types - we first build elements, for example
GenericFunctionTypeElement, and set the type of the GenericFunctionType
nodes to a lazy FunctionType on top of not-yet-resolved element.
Then on the next step TypeResolverVisitor goes over the AST second
time and updates elements with actual types for TypeName(s).

There is no need for doing this now when we are using summary2, it
has its own way to solve forward references with creating elements for
unresolved AST nodes, and breaking cycles with TypeBuilder(s). But the
current implementation is quite big, and it will take some time for me
to try one or another approach to restructure it.

Change-Id: Ic0b8594a6723f715bc5484f9946324c46e30ffb2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/120620
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>