[vm/compiler] Use environment from branch instruction in slow path of CheckedSmiComparison.
The environment of the comparison instruction gets removed when the
branch becomes the owner of the comparison. To ensure the slow path
of CheckedSmiComparison has an environment for generating of catch
entries, we use the environment from the branch.

Fixes https://github.com/dart-lang/sdk/issues/38700

Change-Id: Ib8f86b2b9cc65f606654a89846fabe57a4023cd9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/120587
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Reviewed-by: Aart Bik <ajcbik@google.com>