[vm/ffi] Handle IntConverter's similar to LoadUntagged in load optimizer
Our LoadIndexedInstr/StoreIndexedInstr can operate on an untagged array,
for which we assume it can alias with anything (despite constant
index).

Optimized dart:ffi code will create LoadIndexedInstr/StoreIndexedInstr
that operate on IntConverterInstr. Those can be views into arbitrary
memory so we should assume it can alias with anything (despite constant
index).

Right now it happens to be not a problem (and we cannot have a Dart test
to trigger it), because every LoadIndexedInstr/StoreIndexedInstr that

dart:ffi code generates happens to use a 0 index: It will load the
address of the pointer, adds the scaled index to it and then converts
the result of this arithmetic to an untagged pointer, used by the
load/store with index 0.
Change-Id: I94268486603b78e4d21c3dafaccd133da35193f3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/121843
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>