Update to latest package:yaml
Closes #38958

Tweak an analyzer test that fails after the update. There is a change in
package:yaml where map keys are now iterated in the order they are
defined in the file rather than an arbitrary order based on hash codes.

It is not clear to me whether this is safe or if it indicates a deeper
bug in the analyzer.

Change-Id: Ie02f1bc4277999f991583f9019e67ef70c33a895
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/122029
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>