Analyzer flow analysis for null-aware operators.
The front end implementation is in a previous CL
(https://dart-review.googlesource.com/c/sdk/+/122415).  This CL
contains the analyzer implementation and the integration tests.

Note that the analyzer implementation of null shorting is incomplete,
so some corner cases couldn't be tested.  I'll be fixing this in
follow-up CLs.

Change-Id: I0935ee4c67b24f8bf7c70cbb7ec56dd826fbf73a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/122416
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>