Flow analysis: clean up unit tests using custom matchers.
Previously, we made heavy use of operator== in the flow analysis
tests.  But we also tried to verify that the flow analysis engine
didn't compare types with operator==.  This was kludgy, because it
meant that we had to use a global variable to keep track of whether a
use of _Type.operator== should succeed.  It was also kludgy because it
required an implementation of operator== in flow analysis code that
was solely meant for testing.

By using custom matchers, we can get the same level of testing
flexibility without making use of operator== anywhere.

Also, fixed a few flaws in the tests for "join".  They were
accidentally setting up variables in the already-written state, so the
"assigned" subtest was passing in spite of incorrect expectations.

Change-Id: If326678aeb20a62433ad090968e150738ba5dc59
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/123140
Reviewed-by: Johnni Winther <johnniwinther@google.com>