Use includedElementNames to choose between server and client suggestions
In the case where model doesn't find a CompletionSuggestion or
IncludedSuggestionRelevanceTag we previously created both in case
the client was holding a matching token in its cache. Now we instead
apply a contains check to includedElementNames and will create
an IncludedSuggestionRelevanceTag if a token exists client-side.
Otherwise we will create a CompletionSuggestion.

Bug: https://github.com/dart-lang/sdk/issues/37797
Change-Id: I558734b65d38d8833c414a4c2cc584d0a6d15b28
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/123321
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Ari Aye <ariaye@google.com>