[frontend_server] write JavaScript source maps
Write JavaScript source map contents as part of the incremental compile. places them in a single file adjacent to the source and offset manifests.

This file contains a map where the module name corresponds to the key and the value is the source map object.

Change-Id: Ice8ac910a5d7119a51d567b8f14cef560cfe8e2e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/123360
Reviewed-by: Jake Macdonald <jakemac@google.com>
Commit-Queue: Jonah Williams <jonahwilliams@google.com>